ECS SDK Auto Released By yuanping.wyp,Version：1.29.6

+++ services/ecs/describe_available_resource.go
@@ -77,7 +77,7 @@ func (client *Client) DescribeAvailableResourceWithCallback(request *DescribeAva
 type DescribeAvailableResourceRequest struct {
 	*requests.RpcRequest
 	ResourceOwnerId      requests.Integer `position:"Query" name:"ResourceOwnerId"`
-	Memory               requests.Integer `position:"Query" name:"Memory"`
+	Memory               requests.Float   `position:"Query" name:"Memory"`
 	IoOptimized          string           `position:"Query" name:"IoOptimized"`
 	DataDiskCategory     string           `position:"Query" name:"DataDiskCategory"`
 	Cores                requests.Integer `position:"Query" name:"Cores"`
