+++ services/ecs/describe_resources_modification.go
@@ -77,7 +77,7 @@ func (client *Client) DescribeResourcesModificationWithCallback(request *Describ
 type DescribeResourcesModificationRequest struct {
 	*requests.RpcRequest
 	ResourceOwnerId      requests.Integer `position:"Query" name:"ResourceOwnerId"`
-	Memory               requests.Integer `position:"Query" name:"Memory"`
+	Memory               requests.Float   `position:"Query" name:"Memory"`
 	Cores                requests.Integer `position:"Query" name:"Cores"`
 	MigrateAcrossZone    requests.Boolean `position:"Query" name:"MigrateAcrossZone"`
 	InstanceType         string           `position:"Query" name:"InstanceType"`
